% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tRigon.R
\name{run_tRigon}
\alias{run_tRigon}
\title{Run tRigon application}
\usage{
run_tRigon()
}
\value{
app
}
\description{
Run tRigon application
}
\examples{
if(interactive()){
run_tRigon()
}
}
\author{
David Hoelscher
}
